+++ src/components/calendar/views/weekly-calendar.tsx
@@ -1,8 +1,8 @@
 import { useMemo, useRef } from "react";
+import { useFamilyMembers } from "@/api";
 import { compareEventsByTime, parseTime } from "@/lib/time-utils";
 import { type CalendarEvent, colorMap } from "@/lib/types";
 import { cn } from "@/lib/utils";
-import { useFamilyMembers } from "@/stores";
 import { CalendarEventCard } from "../components/calendar-event";
 import type { FilterState } from "../components/calendar-filter";
 import { CalendarNavigation } from "../components/calendar-navigation";
